README.md
ODIN-RECIPES
this is a recipes webpage i worked on for the odin-project as an assignment on the classes i am taking. this was supposed to be built in html at first by i added little rusty css into it which i learnt throught youtube vids. but now i have upgraded it with new techniques learnt.